Moved generic h5py save/load function to hdf.py
