+++ pycbc/filter/qtransform.py
@@ -45,6 +45,7 @@ import os
 import h5py
 from pycbc.filter import highpass_fir, matched_filter
 from pycbc.psd import welch, interpolate
+from pycbc.io import hdf
 
 def pycbc_insp_main(segments, filename='q_info.hdf'):
     """Main function for pycbc_inspiral implementation of qtransform.py
@@ -127,68 +128,6 @@ def pycbc_insp_tiling(seg, frange=(0,np.inf), qrange=(4,64), mismatch=0.2):
     Qbase, frange = qtiling(data, qrange, frange, mismatch)
 
     return Qbase, frange, data
-
-def save_dict_to_hdf5(dic, filename):
-    """
-    Parameters
-    ----------
-    dic:
-        python dictionary to be converted to hdf5 format
-    filename:
-        desired name of hdf5 file
-    """
-    with h5py.File(filename, 'w') as h5file:
-        recursively_save_dict_contents_to_group(h5file, '/', dic)
-
-def recursively_save_dict_contents_to_group(h5file, path, dic):
-    """
-    Parameters
-    ----------
-    h5file:
-        h5py file to be written to
-    path:
-        path within h5py file to saved dictionary
-    dic:
-        python dictionary to be converted to hdf5 format
-    """
-    for key, item in dic.items():
-        if isinstance(item, (np.ndarray, np.int64, np.float64, str, bytes, tuple, list)):
-            h5file[path + str(key)] = item
-        elif isinstance(item, dict):
-            recursively_save_dict_contents_to_group(h5file, path + key + '/', item)
-        else:
-            raise ValueError('Cannot save %s type'%type(item))
-
-def load_dict_from_hdf5(filename):
-    """
-    Parameters
-    ----------
-    filename:
-        name of h5py file to be loaded
-
-    Returns
-    -------
-    recursively_load_dict_contents_from_group():
-        python dictionary object
-    """
-    with h5py.File(filename, 'r') as h5file:
-        return recursively_load_dict_contents_from_group(h5file, '/')
-
-def recursively_load_dict_contents_from_group(h5file, path):
-    """
-    Parameters
-    ----------
-    h5file:
-        h5py file to be loaded from
-    path:
-        path within h5py file to items
-    """
-    ans = {}
-    for key, item in h5file[path].items():
-        if isinstance(item, h5py._hl.dataset.Dataset):
-            ans[key] = item.value
-
-    
 
 def plotter(interp, out_dir, now, frange, tres, fres):
     """Plotting mechanism for pycbc spectrograms
